Created class for GUI testing with Playwright. Made use of it with the N2 testing code
